@@ -1,7 +1,7 @@
 name: example
 
 environment:
-  sdk: '>=2.0.0-dev.68.0 <3.0.0'
+  sdk: '>=2.12.0-133.2.beta <3.0.0'
 
 dependencies:
   flutter:
