@@ -1,13 +1,12 @@
 name: generic_bloc_provider
 description: A generic BloC Provider for your Flutter apps. This package will help you avoid the boilerplate of writing BloC Providers.
-version: 1.0.9
-author: Roberto Huertas <roberto.huertas@outlook.com>
+version: 2.0.0-nullsafety.0
 homepage: https://github.com/robertohuertasm/generic_bloc_provider
 repository: https://github.com/robertohuertasm/generic_bloc_provider
 issue_tracker: https://github.com/robertohuertasm/generic_bloc_provider/issues
 
 environment:
-  sdk: '>=2.0.0-dev.68.0 <3.0.0'
+  sdk: '>=2.12.0-133.2.beta <3.0.0'
 
 dependencies:
   flutter:
